Change platform_id to review_source_id
`platform_id` could be confused as the ID of a platform, such as
`google`, `facebook`, etc.

What we really mean here is the ID of the review source on the platform.
The review source refers to the Google Business or the Facebook page
from which the review is sourced, hence the name `review_source_id`.