Separate normalization and request logic
Previously normalization and request logic was together in a single
class, which created very long class files that were handling multiple
concerns.

With the introduction of the `Response_Normalizer_Abstract` class, we
can move the normalization logic out of `Request` subclasses. Also, each
platform can now extend it to normalize responses relative to the quirks
of each individual platform.